Upgrading Spring integration test to Jakarta EE 9+.
