+++ http/restful-ws-jakarta-integration-tests/restful-ws-spring/pom.xml
@@ -20,19 +20,19 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>cloudevents-http-restful-ws-integration-tests</artifactId>
+        <artifactId>cloudevents-http-restful-ws-jakarta-integration-tests</artifactId>
         <groupId>io.cloudevents</groupId>
         <version>5.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>cloudevents-http-restful-ws-integration-tests-spring</artifactId>
-    <name>CloudEvents - JAX-RS Integration Tests - Spring</name>
+    <artifactId>cloudevents-http-restful-ws-jakarta-integration-tests-spring</artifactId>
+    <name>CloudEvents - JAX-RS Jakarta EE 9+ Integration Tests - Spring</name>
     <packaging>jar</packaging>
 
     <properties>
         <module-name>io.cloudevents.jaxrs.integration.tests.spring</module-name>
-        <spring-boot.version>2.7.18</spring-boot.version>
+        <spring-boot.version>3.5.14</spring-boot.version>
     </properties>
 
     <dependencyManagement>
@@ -51,7 +51,7 @@
 
         <dependency>
             <groupId>io.cloudevents</groupId>
-            <artifactId>cloudevents-http-restful-ws-integration-tests-common</artifactId>
+            <artifactId>cloudevents-http-restful-ws-jakarta-integration-tests-common</artifactId>
             <version>${project.version}</version>
             <scope>test</scope>
         </dependency>
