+++ http/restful-ws-jakarta-integration-tests/restful-ws-spring/src/test/java/io/cloudevents/http/restful/ws/jakarta/spring/TestApplication.java
@@ -15,16 +15,11 @@
  *
  */
 
-package io.cloudevents.http.restful.ws.spring;
+package io.cloudevents.http.restful.ws.jakarta.spring;
 
-import io.cloudevents.http.restful.ws.CloudEventsProvider;
-import io.cloudevents.http.restful.ws.TestResource;
-import org.glassfish.jersey.server.ResourceConfig;
-import org.springframework.context.annotation.Configuration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
-@Configuration
-public class JerseyConfig extends ResourceConfig {
-    public JerseyConfig() {
-        registerClasses(CloudEventsProvider.class, TestResource.class);
-    }
+@SpringBootApplication(scanBasePackages = {"io.cloudevents.http.restful.ws.jakarta.spring"})
+public class TestApplication extends SpringBootServletInitializer {
 }
